Re-extract if jar file time is changed
Checking the last extraction time is after the jar file time isn't
enough for development version.
We have issues where we first build multiple dev versions and then
test/deploy them a day after. In this case, the cache won't be cleaned
up and the app classpath may be wrong.
Now if you change the jar, it re-extracts.

Change-Id: I9648e2ef7b7fa031b8455465c6bc605fc1c71e72